requirements.txt
flask==2.3.3
requests==2.31.0
plotly==5.17.0
pandas==2.1.4
dash==2.14.2
dash-bootstrap-components==1.5.0
python-dotenv==1.0.0
web3==6.11.3
cryptography==41.0.7
numpy>=1.22.4,<2.0
